requirements.txt
# yolo and label studios
ultralytics
label-studio
pymupdf

torch
torchvision
torchaudio

seaborn
pandas
opencv-python
numpy

# ocr
pytesseract
easyocr
paddlepaddle 
paddleocr

python-dotenv
matplotlib


# langchain
langchain
langchain-core
langchain-google-genai
langchain-community
langchain-openai

PyPDF2
imutils
transformers